refactor addon for later reuse code in CLI
