+++ pkg/apiserver/rest/apis/v1/types.go
@@ -19,10 +19,11 @@ package v1
 import (
 	"time"
 
+	"github.com/oam-dev/kubevela/pkg/addon"
+
 	"github.com/getkin/kin-openapi/openapi3"
 
 	"github.com/oam-dev/kubevela/apis/core.oam.dev/common"
-	"github.com/oam-dev/kubevela/apis/core.oam.dev/v1beta1"
 	"github.com/oam-dev/kubevela/apis/types"
 	"github.com/oam-dev/kubevela/pkg/apiserver/model"
 	"github.com/oam-dev/kubevela/pkg/apiserver/rest/utils"
@@ -56,17 +57,18 @@ type EmptyResponse struct{}
 // CreateAddonRegistryRequest defines the format for addon registry create request
 type CreateAddonRegistryRequest struct {
 	Name string                `json:"name" validate:"checkname"`
-	Git  *model.GitAddonSource `json:"git,omitempty" validate:"required"`
+	Git  *addon.GitAddonSource `json:"git,omitempty" validate:"required"`
 }
 
+// UpdateAddonRegistryRequest defines the format for addon registry update request
 type UpdateAddonRegistryRequest struct {
-	Git *model.GitAddonSource `json:"git,omitempty" validate:"required"`
+	Git *addon.GitAddonSource `json:"git,omitempty" validate:"required"`
 }
 
 // AddonRegistryMeta defines the format for a single addon registry
 type AddonRegistryMeta struct {
 	Name string                `json:"name" validate:"required"`
-	Git  *model.GitAddonSource `json:"git,omitempty"`
+	Git  *addon.GitAddonSource `json:"git,omitempty"`
 }
 
 // ListAddonRegistryResponse list addon registry
@@ -82,53 +84,18 @@ type EnableAddonRequest struct {
 
 // ListAddonResponse defines the format for addon list response
 type ListAddonResponse struct {
-	Addons []*AddonMeta `json:"addons"`
-}
-
-// AddonDeployTo defines where the addon to deploy to
-type AddonDeployTo struct {
-	ControlPlane   bool `json:"control_plane"`
-	RuntimeCluster bool `json:"runtime_cluster"`
-}
-
-// AddonDependency defines the other addons it depends on
-type AddonDependency struct {
-	Name string `json:"name,omitempty"`
-}
-
-// AddonMeta defines the format for a single addon
-type AddonMeta struct {
-	Name         string             `json:"name" validate:"required"`
-	Version      string             `json:"version"`
-	Description  string             `json:"description"`
-	Icon         string             `json:"icon"`
-	URL          string             `json:"url,omitempty"`
-	Tags         []string           `json:"tags,omitempty"`
-	DeployTo     *AddonDeployTo     `json:"deploy_to,omitempty"`
-	Dependencies []*AddonDependency `json:"dependencies,omitempty"`
-}
-
-// AddonElementFile can be addon's definition or addon's component
-type AddonElementFile struct {
-	Data string
-	Name string
-	Path []string
+	Addons []*types.AddonMeta `json:"addons"`
 }
 
 // DetailAddonResponse defines the format for showing the addon details
 type DetailAddonResponse struct {
-	AddonMeta
+	types.AddonMeta
 
 	APISchema *openapi3.Schema     `json:"schema"`
 	UISchema  []*utils.UIParameter `json:"uiSchema"`
 
 	// More details about the addon, e.g. README
-	Detail        string               `json:"detail,omitempty"`
-	Definitions   []AddonElementFile   `json:"definitions"`
-	Parameters    string               `json:"parameters"`
-	CUETemplates  []AddonElementFile   `json:"cue_templates"`
-	YAMLTemplates []AddonElementFile   `json:"yaml_templates,omitempty"`
-	AppTemplate   *v1beta1.Application `json:"app_template"`
+	Detail string `json:"detail,omitempty"`
 }
 
 // AddonStatusResponse defines the format of addon status response
