+++ pkg/apiserver/rest/webservice/addon.go
@@ -20,6 +20,8 @@ import (
 	restfulspec "github.com/emicklei/go-restful-openapi/v2"
 	"github.com/emicklei/go-restful/v3"
 
+	"github.com/oam-dev/kubevela/apis/types"
+
 	apis "github.com/oam-dev/kubevela/pkg/apiserver/rest/apis/v1"
 	"github.com/oam-dev/kubevela/pkg/apiserver/rest/usecase"
 	"github.com/oam-dev/kubevela/pkg/apiserver/rest/utils/bcode"
@@ -103,7 +105,7 @@ func (s *addonWebService) listAddons(req *restful.Request, res *restful.Response
 		return
 	}
 
-	var addons []*apis.AddonMeta
+	var addons []*types.AddonMeta
 
 	for _, d := range detailAddons {
 		addons = append(addons, &d.AddonMeta)
